vakten/vaktpj/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
confusable-homoglyphs==3.2.0
Django==3.1.3
django-ads==1.1.1
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-braces==1.14.0
django-classy-tags==2.0.0
django-dynamic-preferences==1.10.1
django-js-reverse==0.9.1
django-registration==3.1.1
django-sekizai==2.0.0
gunicorn==20.0.4
importlib-metadata==2.1.1
persisting-theory==0.2.1
Pillow==8.0.1

psycopg2-binary==2.8.6
pytz==2020.4
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
zipp==3.4.0
